requirements.txt
absl-py==1.0.0
appdirs==1.4.4
asgiref==3.4.1
astunparse==1.6.3
audioread==2.1.9
beautifulsoup4==4.10.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cycler==0.11.0
decorator==5.1.0
Django==3.2.9
djangorestframework==3.12.4
docopt==0.6.2
flatbuffers==2.0
fonttools==4.28.3
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.8.2
joblib==1.1.0
Js2Py==0.71
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
librosa==0.8.1
llvmlite==0.37.0
Markdown==3.3.6
matplotlib==3.5.1
numba==0.54.1
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
Pillow==8.4.0
pipwin==0.5.1
pooch==1.5.2
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio @ file:///C:/Users/Aryan/pipwin/PyAudio-0.2.11-cp39-cp39-win_amd64.whl
pycparser==2.21
pyjsparser==2.7.1
pyparsing==3.0.6
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.8
scikit-learn==1.0.1
scipy==1.7.3
six==1.16.0
sklearn==0.0
SoundFile==0.10.3.post1
soupsieve==2.3.1
SpeechRecognition==3.8.1
sqlparse==0.4.2
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
threadpoolctl==3.0.0
typing-extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0
